Move Keyboards module to lib/
