+++ keyboard/src/Makefile.am
@@ -12,9 +12,6 @@ schemafilesdir = $(schemadir)/autoyast/rnc
 schemafiles_DATA = \
   autoyast-rnc/keyboard.rnc
 
-ydata_DATA = \
-  data/keyboards.rb
-
 desktop_DATA = \
   desktop/org.opensuse.yast.Keyboard.desktop
 
@@ -25,7 +22,8 @@ ywidgets_DATA = \
 ylibdir = @ylibdir@/y2keyboard
 ylib_DATA = \
   lib/y2keyboard/keyboard_layout_loader.rb \
-  lib/y2keyboard/keyboard_layout.rb
+  lib/y2keyboard/keyboard_layout.rb \
+  lib/y2keyboard/keyboards.rb
 
 ylibclientsdir = @ylibdir@/y2keyboard/clients
 ylibclients_DATA = \
@@ -51,8 +49,8 @@ scalable_DATA = \
   icons/hicolor/scalable/apps/yast-keyboard.svg
 
 EXTRA_DIST = $(module_DATA) $(client_DATA) $(ynclude_DATA) $(scrconf_DATA) \
-  $(schemafiles_DATA) $(ydata_DATA) $(desktop_DATA) $(ywidgets_DATA) \
-  $(ylib_DATA) $(ylibclients_DATA) $(ylibdialogs_DATA) $(ylibstrategies_DATA) \
+  $(schemafiles_DATA) $(desktop_DATA) $(ywidgets_DATA) $(ylib_DATA) \
+  $(ylibclients_DATA) $(ylibdialogs_DATA) $(ylibstrategies_DATA) \
   $(symbolic_DATA) $(scalable_DATA)
 
 include $(top_srcdir)/Makefile.am.common
