+++ keyboard/src/lib/y2keyboard/keyboards.rb
