+++ package/yast2-country.spec
@@ -120,7 +120,6 @@ mkdir -p %{buildroot}%{_datadir}/polkit-1/actions
 %files data
 %dir %{yast_ydatadir}/languages
 %{yast_ydatadir}/languages/*.ycp
-%{yast_ydatadir}/keyboards.rb
 %{yast_moduledir}/Language.rb
 %dir %{yast_libdir}/y2country
 %{yast_libdir}/y2country/language_dbus.rb
